+++ TableSettingsMVC/Resources/Assets.xcassets/Contents.json
