+++ TableSettingsMVC/Resources/Info.plist
